+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 - **[Feature]** Introduce per-topic data exploration in the Explorer.
 - **[Feature]** Introduce per-topic and per-partition most recent message view with live reload.
 - **[Feature]** Introduce a new per-process inspection view called "Details" ti display all process real-time aggregated data.
+- [Improvement] Link all explorable offsets to the explorer.
 - [Improvement] Extend status page checks to ensure, that it detects a case when Web-UI is not part of `karafka.rb` but still referenced in routes.
+- [Improvement] Extend status page checks to ensure, that it detects a case where there is no initial consumers metrics in Kafka topic.
 - [Improvement] Report Rails version when viewing status page (if Rails used).
 - [Improvement] List Web UI topics names on the status page in the info section.
 - [Improvement] Start versioning the materialized states schemas.